requirements.txt
torch
django
djangorestframework